style: move function remove file
